+++ empire/server/api/v2/plugin/plugin_api.py
@@ -2,7 +2,7 @@ from fastapi import Depends, HTTPException
 from sqlalchemy.orm import Session
 
 from empire.server.api.api_router import APIRouter
-from empire.server.api.jwt_auth import get_current_active_user
+from empire.server.api.jwt_auth import get_current_active_user, get_current_user
 from empire.server.api.v2.plugin.plugin_dto import (
     PluginExecutePostRequest,
     PluginExecuteResponse,
@@ -11,6 +11,7 @@ from empire.server.api.v2.plugin.plugin_dto import (
 )
 from empire.server.api.v2.shared_dependencies import get_db
 from empire.server.api.v2.shared_dto import BadRequestResponse, NotFoundResponse
+from empire.server.core.db import models
 from empire.server.server import main
 
 plugin_service = main.pluginsv2
@@ -57,10 +58,12 @@ async def execute_plugin(
     plugin_req: PluginExecutePostRequest,
     plugin=Depends(get_plugin),
     db: Session = Depends(get_db),
+    current_user: models.User = Depends(get_current_user),
 ):
-    results, err = plugin_service.execute_plugin(db, plugin, plugin_req)
+    results, err = plugin_service.execute_plugin(db, plugin, plugin_req, current_user)
 
-    # A plugin can return False for some internal error, or it can raise an actual exception.
+    # A plugin can return False for some internal error,
+    #  or it can raise an actual exception.
     if results is False:
         raise HTTPException(500, "internal plugin error")
     if err: